LayoutTests/imported/w3c/web-platform-tests/trusted-types/csp-block-eval-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.


PASS eval with plain string throws (both block).
FAIL eval with TrustedScript throws (script-src blocks). assert_throws_js: function "_ => {
      eval(p.createScript('a="Hello transformed string"'));
    }" did not throw
